requirements.txt
streamlit
opencv-python-headless

stqdm
streamlit-code-editor